보스몬스터 기본적인 그로기패턴 추가 및, SetFSM 코드에서 TEnableIf대신 if constexpr을 사용해서 코드작성. 더 직관적이라 코드가독성이 좋다.
그로기패턴이 끝나면 스테미나 100%회복 구현.

+++ Source/HyobinsProject/private/Monster/Minions/Super/SuperMinion.cpp
@@ -40,7 +40,7 @@ void ASuperMinion::Activate()
 	Super::Activate();
 
 	GetWorld()->GetGameInstance()->GetSubsystem<UUIManager>()->SetVisibilityWidgets("MonsterHPBar",this, ESlateVisibility::HitTestInvisible);
-	SetFSMState(static_cast<uint8>(ESuperMinionFSMStates::Patrol));
+	SetFSMState(ESuperMinionFSMStates::Patrol);
 }
 
 void ASuperMinion::Tick(float DeltaTime)
